feat: implement analytics Elasticsearch initialization and configuration in the respective modules
